# roc-models/sra-0.2.x/openapi.yaml
...
components:
  parameters:
    store-id:
      content:
        text/plain; charset=utf-8:
          schema:
            type: string
      description: store-id (target in onos-config)
      in: path
      name: store-id
      required: true
...
  schemas:
...
    AdditionalPropertyStore-Id:
      description: Optionally specify a store-id other than the default (only on PATCH
        method)
      properties:
        store-id:
          description: an override of the store-id (target)
          title: store-id
          type: string
      title: AdditionalPropertyStore-Id
      type: object
    AdditionalPropertyUnchanged:
      description: To optionally omit 'required' properties, add them to 'unchanged'
        list
      properties:
        unchanged:
          description: A comma seperated list of unchanged mandatory attribute names
          title: unchanged
          type: string
      title: AdditionalPropertyUnchanged
      type: object
...
    Retail-area:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyStore-Id'
      description: Area is a physical place in a retail location e.g. department,
        floor (single)
      properties:
        area-id:
          description: An identifier for the retail area
          maxLength: 80
          minLength: 1
          title: area-id
          type: string
          x-go-type: ListKey
        description:
          description: long description field
          maxLength: 1024
          minLength: 1
          title: description
          type: string
        display-name:
          description: display name to use in GUI or CLI
          maxLength: 80
          minLength: 1
          title: display-name
          type: string
        image:
          description: a URL to an image file (e.g. SVG, PNG). Optional
          maxLength: 200
          minLength: 1
          title: image
          type: string
        location:
          $ref: '#/components/schemas/Retail-area_Location'
        source:
          $ref: '#/components/schemas/Retail-area_Source_List'
      required:
      - area-id
      title: Retail-area
      type: object
      x-list-multiple: true
    Retail-area_List:
      description: Area is a physical place in a retail location e.g. department,
        floor (list)
      items:
        $ref: '#/components/schemas/Retail-area'
      type: array
      uniqueItems: true
      x-keys:
      - area-id
      x-list-multiple: true
    Retail-area_Location:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyUnchanged'
      description: physical location in grid or latitude and longitude
      properties:
        coordinate-system:
          description: choice of coordinate system
          enum:
          - grid-bottom-left
          - grid-top-left
          - lat-long
          title: coordinate-system
          type: string
        orientation-x:
          description: rotation in dgrees the X-Y plane clockwise from the 3 o'clock
            position
          format: int64
          title: orientation-x
          type: integer
        x:
          description: The X position in grid or longitude by 10e-7
          format: int64
          title: x
          type: integer
        "y":
          description: The Y position in grid or latitude by 10e-7
          format: int64
          title: "y"
          type: integer
        z:
          description: The Z position in grid or elevation in centimetres (optional)
          format: int64
          title: z
          type: integer
      required:
      - coordinate-system
      - x
      - "y"
      title: Retail-area_Location
      type: object
    Retail-area_Source:
      description: The list of cameras (single)
      properties:
        description:
          description: long description field
          maxLength: 1024
          minLength: 1
          title: description
          type: string
        display-name:
          description: display name to use in GUI or CLI
          maxLength: 80
          minLength: 1
          title: display-name
          type: string
        image:
          description: a URL to an image file (e.g. SVG, PNG). Optional
          maxLength: 200
          minLength: 1
          title: image
          type: string
        location:
          $ref: '#/components/schemas/Retail-area_Source_Location'
        source-id:
          description: the name of the inference source
          title: source-id
          type: string
          x-go-type: ListKey
        state:
          $ref: '#/components/schemas/Retail-area_Source_State'
        video:
          $ref: '#/components/schemas/Retail-area_Source_Video'
      required:
      - source-id
      title: Retail-area_Source
      type: object
      x-list-multiple: true
    Retail-area_Source_List:
      description: The list of cameras (list)
      items:
        $ref: '#/components/schemas/Retail-area_Source'
      type: array
      uniqueItems: true
      x-keys:
      - source-id
      x-list-multiple: true
    Retail-area_Source_Location:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyUnchanged'
      description: physical location in grid or latitude and longitude
      properties:
        coordinate-system:
          description: choice of coordinate system
          enum:
          - grid-bottom-left
          - grid-top-left
          - lat-long
          title: coordinate-system
          type: string
        orientation-x:
          description: rotation in dgrees the X-Y plane clockwise from the 3 o'clock
            position
          format: int64
          title: orientation-x
          type: integer
        x:
          description: The X position in grid or longitude by 10e-7
          format: int64
          title: x
          type: integer
        "y":
          description: The Y position in grid or latitude by 10e-7
          format: int64
          title: "y"
          type: integer
        z:
          description: The Z position in grid or elevation in centimetres (optional)
          format: int64
          title: z
          type: integer
      required:
      - coordinate-system
      - x
      - "y"
      title: Retail-area_Source_Location
      type: object
    Retail-area_Source_State:
      description: read only attributes of the source - populated by adapter from
        analytics engine
      properties:
        error:
          description: description of error if present
          readOnly: true
          title: error
          type: string
        error-since:
          description: date and time error last updated
          readOnly: true
          title: error-since
          type: string
        status:
          description: status of the source
          readOnly: true
          title: status
          type: string
      title: Retail-area_Source_State
      type: object
    Retail-area_Source_Video:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyUnchanged'
      description: attributes related to video stream
      properties:
        path:
          description: the path to the inference source
          maxLength: 100
          minLength: 1
          pattern: '[a-zA-Z0-9\./_:-]*'
          title: path
          type: string
        source-type:
          description: Source type
          enum:
          - device
          - file
          - sample
          - stream
          title: source-type
          type: string
      required:
      - path
      - source-type
      title: Retail-area_Source_Video
      type: object
...
info:
  contact:
    name: Intel Corporation
    url: https://www.intel.com
  description: OpenAPI 3 specification is generated from sra onos-config model plugin
  license:
    name: LicenseRef-Intel
  title: sra-0.2.x
  version: 0.2.x
openapi: 3.0.0
paths:
  /sra/v0.2.x/{store-id}/retail-area:
    description: Area is a physical place in a retail location e.g. department, floor
    get:
      operationId: getRetail-area_List
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Retail-area_List'
          description: GET OK 200
      summary: GET /retail-area List
      tags:
      - Retail-area
    parameters:
    - $ref: '#/components/parameters/store-id'
...